requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2024.7.4
charset-normalizer==3.3.2
DateTime==5.4
exchange_calendars==4.5.3
finvizfinance==1.1.0
frozendict==2.4.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
html5lib==1.1
httplib2==0.22.0
idna==3.7
korean-lunar-calendar==0.3.1
lxml==5.1.0
multitasking==0.0.11
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.1
pandas_market_calendars==4.4.0
peewee==3.17.1
protobuf==4.25.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyluach==2.2.0
pyparsing==3.1.1
pypdf==5.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.4
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
toolz==0.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.5.0
webencodings==0.5.1
yfinance==0.2.61
zope.interface==6.2
